generate istio manifest
