+++ bootstrap/pkg/apis/apps/ksonnet/v1alpha1/application_types.go
@@ -188,12 +188,12 @@ type RegistriesConfigFile struct {
 }
 
 type AppConfig struct {
-	Registries      []*RegistryConfig `json:"registries,omitempty"`
-	Packages        []KsPackage       `json:"packages,omitempty"`
-	Components      []KsComponent     `json:"components,omitempty"`
-	Parameters      []KsParameter     `json:"parameters,omitempty"`
+	Registries []*RegistryConfig `json:"registries,omitempty"`
+	Packages   []KsPackage       `json:"packages,omitempty"`
+	Components []KsComponent     `json:"components,omitempty"`
+	Parameters []KsParameter     `json:"parameters,omitempty"`
 	// Parameters to apply when creating the ksonnet components
-	ApplyParameters []KsParameter     `json:"applyParameters,omitempty"`
+	ApplyParameters []KsParameter `json:"applyParameters,omitempty"`
 }
 
 type NameValue struct {
@@ -252,7 +252,7 @@ type Ksonnet struct {
 type KsAppList struct {
 	metav1.TypeMeta `json:",inline"`
 	metav1.ListMeta `json:"metadata,omitempty"`
-	Items []Ksonnet `json:"items"`
+	Items           []Ksonnet `json:"items"`
 }
 
 func init() {
