+++ bootstrap/pkg/client/gcp/gcp.go
@@ -17,40 +17,30 @@ limitations under the License.
 package gcp
 
 import (
-	"encoding/base64"
-
 	"fmt"
+	"io"
+	"io/ioutil"
+	"net/http"
+	"os"
+	"path"
+	"path/filepath"
+	"regexp"
+
 	"github.com/cenkalti/backoff"
 	"github.com/ghodss/yaml"
-	gogetter "github.com/hashicorp/go-getter"
 	configtypes "github.com/kubeflow/kubeflow/bootstrap/config"
 	kftypes "github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps"
 	gcptypes "github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps/gcp/v1alpha1"
-	kstypes "github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps/ksonnet/v1alpha1"
 	"github.com/kubeflow/kubeflow/bootstrap/pkg/client/ksonnet"
-	kfctlutils "github.com/kubeflow/kubeflow/bootstrap/pkg/utils"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/net/context"
-	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/google"
 	gke "google.golang.org/api/container/v1"
 	"google.golang.org/api/deploymentmanager/v2"
-	"google.golang.org/api/iam/v1"
-	"google.golang.org/api/serviceusage/v1"
-	"io"
-	"io/ioutil"
 	v1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientset "k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-	"net/http"
-	"os"
-	"os/exec"
-	"path"
-	"path/filepath"
-	"regexp"
-	"strings"
 )
 
 const (
@@ -60,6 +50,9 @@ const (
 	STORAGE_FILE      = "storage-kubeflow.yaml"
 	NETWORK_FILE      = "network.yaml"
 	GCFS_FILE         = "gcfs.yaml"
+	ISTIO_DIR         = "istio"
+	ISTIO_CRD         = "istio-crd.yaml"
+	ISTIO_INSTALL     = "istio-noauth.yaml"
 	ADMIN_SECRET_NAME = "admin-gcp-sa"
 	USER_SECRET_NAME  = "user-gcp-sa"
 	KUBEFLOW_OAUTH    = "kubeflow-oauth"
@@ -163,11 +156,44 @@ func GetKfApp(options map[string]interface{}) kftypes.KfApp {
 	if options[string(kftypes.USE_BASIC_AUTH)] != nil {
 		_gcp.GcpApp.Spec.UseBasicAuth = options[string(kftypes.USE_BASIC_AUTH)].(bool)
 	}
+	if options[string(kftypes.USE_ISTIO)] != nil {
+		_gcp.GcpApp.Spec.UseIstio = options[string(kftypes.USE_ISTIO)].(bool)
+	}
 	if options[string(kftypes.SKIP_INIT_GCP_PROJECT)] != nil {
 		skipInitProject := options[string(kftypes.SKIP_INIT_GCP_PROJECT)].(bool)
 		_gcp.GcpApp.Spec.SkipInitProject = skipInitProject
 	}
 	return _gcp
+}
+
+// Delete the gcp kfapp
+func (gcp *Gcp) Delete(resources kftypes.ResourceEnum, options map[string]interface{}) error {
+	ks := gcp.Children[kftypes.KSONNET]
+	if ks != nil {
+		ksDeleteErr := ks.Delete(resources, options)
+		if ksDeleteErr != nil {
+			return fmt.Errorf("gcp delete failed for %v: %v", string(kftypes.KSONNET), ksDeleteErr)
+		}
+	} else {
+		return fmt.Errorf("%v not in Children", string(kftypes.KSONNET))
+	}
+	return nil
+}
+
+func (gcp *Gcp) ConfigK8s() error {
+	ctx := context.Background()
+	k8sClientset, err := gcp.getK8sClientset(ctx)
+	if err != nil {
+		return err
+	}
+	if err = createNamespace(k8sClientset, gcp.GcpApp.Namespace); err != nil {
+		return fmt.Errorf("Creating namespace error: %v", err)
+	}
+	if err = bindAdmin(k8sClientset, gcp.GcpApp.Spec.Email); err != nil {
+		return fmt.Errorf("Binding user as admin error: %v", err)
+	}
+
+	return nil
 }
 
 func getSA(name string, nameSuffix string, project string) string {
@@ -279,47 +305,6 @@ func blockingWait(project string, opName string, deploymentmanagerService *deplo
 	}, backoff.NewExponentialBackOff())
 }
 
-func (gcp *Gcp) updateDeployment(deployment string, yamlfile string) error {
-	appDir := gcp.GcpApp.Spec.AppDir
-	gcpConfigDir := path.Join(appDir, GCP_CONFIG)
-	ctx := context.Background()
-	client, clientErr := google.DefaultClient(ctx, deploymentmanager.CloudPlatformScope)
-	if clientErr != nil {
-		return fmt.Errorf("Error getting DefaultClient: %v", clientErr)
-	}
-	deploymentmanagerService, err := deploymentmanager.New(client)
-	if err != nil {
-		return fmt.Errorf("Error creating deploymentmanagerService: %v", err)
-	}
-	filePath := filepath.Join(gcpConfigDir, yamlfile)
-	dp := &deploymentmanager.Deployment{
-		Name: deployment,
-	}
-	if target, targetErr := generateTarget(filePath); targetErr != nil {
-		return targetErr
-	} else {
-		dp.Target = target
-	}
-
-	project := gcp.GcpApp.Spec.Project
-	resp, err := deploymentmanagerService.Deployments.Get(project, deployment).Context(ctx).Do()
-	if err == nil {
-		dp.Fingerprint = resp.Fingerprint
-		op, updateErr := deploymentmanagerService.Deployments.Update(project, deployment, dp).Context(ctx).Do()
-		if updateErr != nil {
-			return fmt.Errorf("Update deployment error: %v", updateErr)
-		}
-		return blockingWait(project, op.Name, deploymentmanagerService, ctx)
-	} else {
-		log.Infof("Get deployment error, creating: %v", err)
-		op, insertErr := deploymentmanagerService.Deployments.Insert(project, dp).Context(ctx).Do()
-		if insertErr != nil {
-			return fmt.Errorf("Insert deployment error: %v", insertErr)
-		}
-		return blockingWait(project, op.Name, deploymentmanagerService, ctx)
-	}
-}
-
 func createNamespace(k8sClientset *clientset.Clientset, namespace string) error {
 	log.Infof("Creating namespace: %v", namespace)
 	_, err := k8sClientset.CoreV1().Namespaces().Get(namespace, metav1.GetOptions{})
@@ -379,140 +364,6 @@ func bindAdmin(k8sClientset *clientset.Clientset, user string) error {
 	return err
 }
 
-func (gcp *Gcp) ConfigK8s() error {
-	ctx := context.Background()
-	k8sClientset, err := gcp.getK8sClientset(ctx)
-	if err != nil {
-		return err
-	}
-	if err = createNamespace(k8sClientset, gcp.GcpApp.Namespace); err != nil {
-		return fmt.Errorf("Creating namespace error: %v", err)
-	}
-	if err = bindAdmin(k8sClientset, gcp.GcpApp.Spec.Email); err != nil {
-		return fmt.Errorf("Binding user as admin error: %v", err)
-	}
-
-	return nil
-}
-
-func (gcp *Gcp) updateDM(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	if err := gcp.updateDeployment(gcp.GcpApp.Name+"-storage", STORAGE_FILE); err != nil {
-		return fmt.Errorf("could not update %v: %v", STORAGE_FILE, err)
-	}
-	if err := gcp.updateDeployment(gcp.GcpApp.Name, CONFIG_FILE); err != nil {
-		return fmt.Errorf("could not update %v: %v", CONFIG_FILE, err)
-	}
-	if _, networkStatErr := os.Stat(path.Join(gcp.GcpApp.Spec.AppDir, NETWORK_FILE)); !os.IsNotExist(networkStatErr) {
-		err := gcp.updateDeployment(gcp.GcpApp.Name+"-network", NETWORK_FILE)
-		if err != nil {
-			return fmt.Errorf("could not update %v: %v", NETWORK_FILE, err)
-		}
-	}
-	if _, gcfsStatErr := os.Stat(path.Join(gcp.GcpApp.Spec.AppDir, GCFS_FILE)); !os.IsNotExist(gcfsStatErr) {
-		err := gcp.updateDeployment(gcp.GcpApp.Name+"-gcfs", GCFS_FILE)
-		if err != nil {
-			return fmt.Errorf("could not update %v: %v", GCFS_FILE, err)
-		}
-	}
-
-	policy, policyErr := kfctlutils.GetIamPolicy(gcp.GcpApp.Spec.Project)
-	if policyErr != nil {
-		return fmt.Errorf("GetIamPolicy error: %v", policyErr)
-	}
-	appDir := gcp.GcpApp.Spec.AppDir
-	gcpConfigDir := path.Join(appDir, GCP_CONFIG)
-	iamPolicy, iamPolicyErr := kfctlutils.ReadIamBindingsYAML(
-		filepath.Join(gcpConfigDir, "iam_bindings.yaml"))
-	if iamPolicyErr != nil {
-		return fmt.Errorf("Read IAM policy YAML error: %v", iamPolicyErr)
-	}
-	kfctlutils.RewriteIamPolicy(policy, iamPolicy, nil)
-	if err := kfctlutils.SetIamPolicy(gcp.GcpApp.Spec.Project, policy); err != nil {
-		return fmt.Errorf("SetIamPolicy error: %v", err)
-	}
-
-	if err := gcp.ConfigK8s(); err != nil {
-		return fmt.Errorf("Configure K8s is failed: %v", err)
-	}
-
-	ctx := context.Background()
-	cluster, err := kftypes.GetClusterInfo(ctx, gcp.GcpApp.Spec.Project,
-		gcp.GcpApp.Spec.Zone, gcp.GcpApp.Name)
-	if err != nil {
-		return fmt.Errorf("Get Cluster error: %v", err)
-	}
-	client, err := kftypes.BuildConfigFromClusterInfo(ctx, cluster)
-	if err != nil {
-		return fmt.Errorf("Build ClientConfig error: %v", err)
-	}
-
-	// TODO(#2604): Need to create a named context.
-	cred_cmd := exec.Command("gcloud", "container", "clusters", "get-credentials",
-		gcp.GcpApp.Name,
-		"--zone="+gcp.GcpApp.Spec.Zone,
-		"--project="+gcp.GcpApp.Spec.Project)
-	cred_cmd.Stdout = os.Stdout
-	log.Infof("Running get-credentials ...")
-	if err = cred_cmd.Run(); err != nil {
-		return fmt.Errorf("Error when running gcloud container clusters get-credentials: %v", err)
-	}
-
-	k8sSpecsDir := path.Join(appDir, K8S_SPECS)
-	daemonsetPreloaded := filepath.Join(k8sSpecsDir, "daemonset-preloaded.yaml")
-	daemonsetPreloadedErr := kfctlutils.RunKubectlApply(daemonsetPreloaded)
-	if daemonsetPreloadedErr != nil {
-		return fmt.Errorf("could not create resources in daemonset-preloaded.yaml %v", daemonsetPreloadedErr)
-	}
-	adminClient := rest.CopyConfig(client)
-	adminClient.Impersonate.UserName = "admin"
-	adminClient.Impersonate.Groups = []string{"system:masters"}
-	rbacSetup := filepath.Join(k8sSpecsDir, "rbac-setup.yaml")
-	rbacSetupErr := kfctlutils.RunKubectlApply(rbacSetup)
-	if rbacSetupErr != nil {
-		return fmt.Errorf("could not create resources in rbac-setup.yaml %v", rbacSetupErr)
-	}
-	agents := filepath.Join(k8sSpecsDir, "agents.yaml")
-	agentsErr := kfctlutils.RunKubectlApply(agents)
-	if agentsErr != nil {
-		return fmt.Errorf("could not create resources in agents.yaml %v", agents)
-	}
-	return nil
-}
-
-func (gcp *Gcp) Apply(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	updateDMErr := gcp.updateDM(resources, options)
-	if updateDMErr != nil {
-		return fmt.Errorf("gcp apply could not update deployment manager Error %v", updateDMErr)
-	}
-	secretsErr := gcp.createSecrets(options)
-	if secretsErr != nil {
-		return fmt.Errorf("gcp apply could not create secrets Error %v", secretsErr)
-	}
-	ks := gcp.Children[kftypes.KSONNET]
-	if ks != nil {
-		ksApplyErr := ks.Apply(resources, options)
-		if ksApplyErr != nil {
-			return fmt.Errorf("gcp apply failed for %v: %v", string(kftypes.KSONNET), ksApplyErr)
-		}
-	} else {
-		return fmt.Errorf("%v not in Children", string(kftypes.KSONNET))
-	}
-	return nil
-}
-
-func (gcp *Gcp) Delete(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	ks := gcp.Children[kftypes.KSONNET]
-	if ks != nil {
-		ksDeleteErr := ks.Delete(resources, options)
-		if ksDeleteErr != nil {
-			return fmt.Errorf("gcp delete failed for %v: %v", string(kftypes.KSONNET), ksDeleteErr)
-		}
-	} else {
-		return fmt.Errorf("%v not in Children", string(kftypes.KSONNET))
-	}
-	return nil
-}
-
 func (gcp *Gcp) copyFile(source string, dest string) error {
 	from, err := os.Open(source)
 	if err != nil {
@@ -547,196 +398,11 @@ func setNameVal(entries []configtypes.NameValue, name string, val string) {
 	})
 }
 
-func (gcp *Gcp) generateKsonnet(options map[string]interface{}) error {
-	configPath := path.Join(gcp.GcpApp.Spec.AppDir,
-		kftypes.DefaultCacheDir,
-		gcp.GcpApp.Spec.Version,
-		kftypes.GcpConfigDir)
-	if gcp.GcpApp.Spec.UseBasicAuth {
-		configPath = path.Join(configPath, kftypes.GcpBasicAuth)
-	} else {
-		configPath = path.Join(configPath, kftypes.GcpIapConfig)
-	}
-	if options[string(kftypes.DefaultConfig)] == nil {
-		options[string(kftypes.DefaultConfig)] = configPath
-	}
-
-	if options[string(kftypes.EMAIL)] != nil &&
-		options[string(kftypes.EMAIL)].(string) != "" {
-		gcp.GcpApp.Spec.Email = options[string(kftypes.EMAIL)].(string)
-	} else if gcp.GcpApp.Spec.Email == "" {
-		return fmt.Errorf("Email is not set in default nor passed.")
-	} else {
-		options[string(kftypes.EMAIL)] = gcp.GcpApp.Spec.Email
-	}
-	if options[string(kftypes.IPNAME)] != nil &&
-		options[string(kftypes.IPNAME)].(string) != "" {
-		gcp.GcpApp.Spec.IpName = options[string(kftypes.IPNAME)].(string)
-	} else if gcp.GcpApp.Spec.IpName == "" {
-		return fmt.Errorf("ipName is not set in default nor passed.")
-	} else {
-		log.Infof("Using default ipName: %v", gcp.GcpApp.Spec.IpName)
-		options[string(kftypes.IPNAME)] = gcp.GcpApp.Spec.IpName
-	}
-
-	if gcp.GcpApp.Spec.UseBasicAuth {
-		options[string(kftypes.USE_BASIC_AUTH)] = true
-	} else {
-		options[string(kftypes.USE_BASIC_AUTH)] = false
-	}
-	if options[string(kftypes.HOSTNAME)] != nil &&
-		options[string(kftypes.HOSTNAME)].(string) != "" {
-		gcp.GcpApp.Spec.Hostname = options[string(kftypes.HOSTNAME)].(string)
-	} else if gcp.GcpApp.Spec.Hostname == "" {
-		return fmt.Errorf("hostname is not set in default nor passed.")
-	} else {
-		log.Infof("Using default hostname: %v", gcp.GcpApp.Spec.Hostname)
-		options[string(kftypes.HOSTNAME)] = gcp.GcpApp.Spec.Hostname
-	}
-	if options[string(kftypes.ZONE)] != nil {
-		gcp.GcpApp.Spec.Zone = options[string(kftypes.ZONE)].(string)
-	}
-	ks := gcp.Children[kftypes.KSONNET]
-	if ks != nil {
-		ksGenerateErr := ks.Generate(kftypes.ALL, options)
-		if ksGenerateErr != nil {
-			return fmt.Errorf("gcp generate failed for %v: %v", string(kftypes.KSONNET), ksGenerateErr)
-		}
-	} else {
-		return fmt.Errorf("%v not in Children", string(kftypes.KSONNET))
-	}
-	return nil
-}
-
 //TODO(#2515)
 func (gcp *Gcp) replaceText(regex string, repl string, src []byte) []byte {
 	re := regexp.MustCompile(regex)
 	buf := re.ReplaceAll(src, []byte(repl))
 	return buf
-}
-
-// TODO(#2515): Switch from string replacement to YAML config.
-func (gcp *Gcp) generateDMConfigs(options map[string]interface{}) error {
-	// TODO(gabrielwen): Use YAML support instead of string replacement.
-	appDir := gcp.GcpApp.Spec.AppDir
-	gcpConfigDir := path.Join(appDir, GCP_CONFIG)
-	gcpConfigDirErr := os.Mkdir(gcpConfigDir, os.ModePerm)
-	if gcpConfigDirErr != nil {
-		return fmt.Errorf("cannot create directory %v", gcpConfigDirErr)
-	}
-	repo := gcp.GcpApp.Spec.Repo
-	parentDir := path.Dir(repo)
-	sourceDir := path.Join(parentDir, "deployment/gke/deployment_manager_configs")
-	files := []string{"cluster-kubeflow.yaml", "cluster.jinja", "cluster.jinja.schema",
-		"storage-kubeflow.yaml", "storage.jinja", "storage.jinja.schema"}
-	for _, file := range files {
-		sourceFile := filepath.Join(sourceDir, file)
-		destFile := filepath.Join(gcpConfigDir, file)
-		copyErr := gcp.copyFile(sourceFile, destFile)
-		if copyErr != nil {
-			return fmt.Errorf("could not copy %v to %v Error %v", sourceFile, destFile, copyErr)
-		}
-	}
-	from := filepath.Join(sourceDir, "iam_bindings_template.yaml")
-	to := filepath.Join(gcpConfigDir, "iam_bindings.yaml")
-	iamBindings := map[string]string{
-		"from": from,
-		"to":   to,
-	}
-	iamBindingsErr := gcp.copyFile(iamBindings["from"], iamBindings["to"])
-	if iamBindingsErr != nil {
-		return fmt.Errorf("could not copy iam_bindings Error %v", iamBindingsErr)
-	}
-	iamBindingsData, iamBindingsDataErr := ioutil.ReadFile(to) // just pass the file name
-	if iamBindingsDataErr != nil {
-		return fmt.Errorf("could not read %v Error %v", to, iamBindingsDataErr)
-	}
-	adminEmail := getSA(gcp.GcpApp.Name, "admin", gcp.GcpApp.Spec.Project)
-	repl := "serviceAccount:" + adminEmail
-	iamBindingsData = gcp.replaceText("set-kubeflow-admin-service-account", repl, iamBindingsData)
-	userEmail := getSA(gcp.GcpApp.Name, "user", gcp.GcpApp.Spec.Project)
-	repl = "serviceAccount:" + userEmail
-	iamBindingsData = gcp.replaceText("set-kubeflow-user-service-account", repl, iamBindingsData)
-	vmEmail := getSA(gcp.GcpApp.Name, "vm", gcp.GcpApp.Spec.Project)
-	repl = "serviceAccount:" + vmEmail
-	iamBindingsData = gcp.replaceText("set-kubeflow-vm-service-account", repl, iamBindingsData)
-	iamEntry := "serviceAccount:" + gcp.GcpApp.Spec.Email
-	re := regexp.MustCompile("iam.gserviceaccount.com")
-	if !re.MatchString(gcp.GcpApp.Spec.Email) {
-		iamEntry = "user:" + gcp.GcpApp.Spec.Email
-	}
-	iamBindingsData = gcp.replaceText("set-kubeflow-iap-account", iamEntry, iamBindingsData)
-	srcErr := ioutil.WriteFile(to, iamBindingsData, 0644)
-	if srcErr != nil {
-		return fmt.Errorf("cound not write to %v Error %v", to, srcErr)
-	}
-	configFile := filepath.Join(gcpConfigDir, CONFIG_FILE)
-	configFileData, configFileDataErr := ioutil.ReadFile(configFile)
-	if configFileDataErr != nil {
-		return fmt.Errorf("could not read %v Error %v", configFile, configFileDataErr)
-	}
-	storageFile := filepath.Join(gcpConfigDir, STORAGE_FILE)
-	storageFileData, storageFileDataErr := ioutil.ReadFile(storageFile)
-	if storageFileDataErr != nil {
-		return fmt.Errorf("could not read %v Error %v", storageFile, storageFileDataErr)
-	}
-	configFileData = gcp.replaceText("SET_GKE_API_VERSION", kftypes.DefaultGkeApiVer, configFileData)
-	repl = "zone: " + gcp.GcpApp.Spec.Zone
-	configFileData = gcp.replaceText("zone: SET_THE_ZONE", repl, configFileData)
-	storageFileData = gcp.replaceText("zone: SET_THE_ZONE", repl, storageFileData)
-	repl = "users: [\"" + iamEntry + "\"]"
-	configFileData = gcp.replaceText("users:", repl, configFileData)
-	repl = "ipName: " + gcp.GcpApp.Spec.IpName
-	configFileData = gcp.replaceText("ipName: kubeflow-ip", repl, configFileData)
-	configFileErr := ioutil.WriteFile(configFile, configFileData, 0644)
-	if configFileErr != nil {
-		return fmt.Errorf("cound not write to %v Error %v", configFile, configFileErr)
-	}
-	repl = "createPipelinePersistentStorage: true"
-	storageFileData = gcp.replaceText("createPipelinePersistentStorage: SET_CREATE_PIPELINE_PERSISTENT_STORAGE",
-		repl, storageFileData)
-	storageFileErr := ioutil.WriteFile(storageFile, storageFileData, 0644)
-	if storageFileErr != nil {
-		return fmt.Errorf("cound not write to %v Error %v", storageFile, storageFileErr)
-	}
-	return nil
-}
-
-func (gcp *Gcp) downloadK8sManifests() error {
-	appDir := gcp.GcpApp.Spec.AppDir
-	k8sSpecsDir := path.Join(appDir, K8S_SPECS)
-	k8sSpecsDirErr := os.Mkdir(k8sSpecsDir, os.ModePerm)
-	if k8sSpecsDirErr != nil {
-		return fmt.Errorf("cannot create directory %v Error %v", k8sSpecsDir, k8sSpecsDirErr)
-	}
-	daemonsetPreloaded := filepath.Join(k8sSpecsDir, "daemonset-preloaded.yaml")
-	url := "https://raw.githubusercontent.com/GoogleCloudPlatform/container-engine-accelerators/stable/nvidia-driver-installer/cos/daemonset-preloaded.yaml"
-	urlErr := gogetter.GetFile(daemonsetPreloaded, url)
-	if urlErr != nil {
-		return fmt.Errorf("couldn't download %v Error %v", url, urlErr)
-	}
-	rbacSetup := filepath.Join(k8sSpecsDir, "rbac-setup.yaml")
-	url = "https://storage.googleapis.com/stackdriver-kubernetes/stable/rbac-setup.yaml"
-	urlErr = gogetter.GetFile(rbacSetup, url)
-	if urlErr != nil {
-		return fmt.Errorf("couldn't download %v Error %v", url, urlErr)
-	}
-	agents := filepath.Join(k8sSpecsDir, "agents.yaml")
-	url = "https://storage.googleapis.com/stackdriver-kubernetes/stable/agents.yaml"
-	urlErr = gogetter.GetFile(agents, url)
-	if urlErr != nil {
-		return fmt.Errorf("couldn't download %v Error %v", url, urlErr)
-	}
-
-	//TODO - copied from scripts/gke/util.sh. The rbac-setup command won't need admin since the user will be
-	// running as admin.
-	//  # Install the GPU driver. It has no effect on non-GPU nodes.
-	//  kubectl apply -f ${KUBEFLOW_K8S_MANIFESTS_DIR}/daemonset-preloaded.yaml
-	//  # Install Stackdriver Kubernetes agents.
-	//  kubectl apply -f ${KUBEFLOW_K8S_MANIFESTS_DIR}/rbac-setup.yaml --as=admin --as-group=system:masters
-	//  kubectl apply -f ${KUBEFLOW_K8S_MANIFESTS_DIR}/agents.yaml
-
-	return nil
 }
 
 func (gcp *Gcp) writeGcpSecret(client *clientset.Clientset, secretName string, data map[string][]byte) error {
@@ -751,137 +417,6 @@ func (gcp *Gcp) writeGcpSecret(client *clientset.Clientset, secretName string, d
 	return err
 }
 
-func (gcp *Gcp) createGcpSecret(ctx context.Context, client *clientset.Clientset,
-	email string, secretName string) error {
-	namespace := gcp.GcpApp.Namespace
-	_, err := client.CoreV1().Secrets(namespace).Get(secretName, metav1.GetOptions{})
-	if err == nil {
-		log.Infof("Secret for %v already exists ...", secretName)
-		return nil
-	}
-
-	log.Infof("Secret for %v not found, creating ...", secretName)
-	ts, err := google.DefaultTokenSource(ctx, iam.CloudPlatformScope)
-	if err != nil {
-		return fmt.Errorf("Get IAM token source error: %v", err)
-	}
-	oClient := oauth2.NewClient(ctx, ts)
-	iamService, err := iam.New(oClient)
-	if err != nil {
-		return fmt.Errorf("Get Oauth client error: %v", err)
-	}
-	name := fmt.Sprintf("projects/%v/serviceAccounts/%v", gcp.GcpApp.Spec.Project,
-		email)
-	req := &iam.CreateServiceAccountKeyRequest{
-		KeyAlgorithm:   "KEY_ALG_RSA_2048",
-		PrivateKeyType: "TYPE_GOOGLE_CREDENTIALS_FILE",
-	}
-	saKey, err := iamService.Projects.ServiceAccounts.Keys.Create(name, req).Context(ctx).Do()
-	if err != nil {
-		return fmt.Errorf("Service account key creation error: %v", err)
-	}
-	privateKeyData, err := base64.StdEncoding.DecodeString(saKey.PrivateKeyData)
-	if err != nil {
-		return fmt.Errorf("PrivateKeyData decoding error: %v", err)
-	}
-	return gcp.writeGcpSecret(client, secretName, map[string][]byte{
-		secretName + ".json": privateKeyData,
-	})
-}
-
-func (gcp *Gcp) createSecrets(options map[string]interface{}) error {
-	ctx := context.Background()
-	k8sClient, err := gcp.getK8sClientset(ctx)
-	if err != nil {
-		return fmt.Errorf("Get K8s clientset error: %v", err)
-	}
-	adminEmail := getSA(gcp.GcpApp.Name, "admin", gcp.GcpApp.Spec.Project)
-	userEmail := getSA(gcp.GcpApp.Name, "user", gcp.GcpApp.Spec.Project)
-	if err := gcp.createGcpSecret(ctx, k8sClient, adminEmail, ADMIN_SECRET_NAME); err != nil {
-		return fmt.Errorf("cannot create admin secret %v Error %v", ADMIN_SECRET_NAME, err)
-
-	}
-	if err := gcp.createGcpSecret(ctx, k8sClient, userEmail, USER_SECRET_NAME); err != nil {
-		return fmt.Errorf("cannot create user secret %v Error %v", USER_SECRET_NAME, err)
-
-	}
-
-	_, err = k8sClient.CoreV1().Secrets(gcp.GcpApp.Namespace).Get(KUBEFLOW_OAUTH, metav1.GetOptions{})
-	if err == nil {
-		log.Infof("Secret for %v already exits ...", KUBEFLOW_OAUTH)
-		return nil
-	}
-
-	oauthId := ""
-	if options[string(kftypes.OAUTH_ID)] != nil &&
-		options[string(kftypes.OAUTH_ID)].(string) != "" {
-		oauthId = options[string(kftypes.OAUTH_ID)].(string)
-	} else {
-		oauthId = os.Getenv(CLIENT_ID)
-	}
-	if oauthId == "" {
-		return fmt.Errorf("At least one of --%v or ENV `%v` needs to be set.",
-			string(kftypes.OAUTH_ID), CLIENT_ID)
-	}
-	oauthSecret := ""
-	if options[string(kftypes.OAUTH_SECRET)] != nil &&
-		options[string(kftypes.OAUTH_SECRET)].(string) != "" {
-		oauthSecret = options[string(kftypes.OAUTH_SECRET)].(string)
-	} else {
-		oauthSecret = os.Getenv(CLIENT_SECRET)
-	}
-	if oauthSecret == "" {
-		return fmt.Errorf("At least one of --%v or ENV `%v` needs to be set.",
-			string(kftypes.OAUTH_SECRET), CLIENT_SECRET)
-	}
-
-	return gcp.writeGcpSecret(k8sClient, KUBEFLOW_OAUTH, map[string][]byte{
-		strings.ToLower(CLIENT_ID):     []byte(oauthId),
-		strings.ToLower(CLIENT_SECRET): []byte(oauthSecret),
-	})
-}
-
-func (gcp *Gcp) Generate(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	switch resources {
-	case kftypes.K8S:
-		generateK8sSpecsErr := gcp.downloadK8sManifests()
-		if generateK8sSpecsErr != nil {
-			return fmt.Errorf("could not generate files under %v Error: %v", K8S_SPECS, generateK8sSpecsErr)
-		}
-		ksonnetErr := gcp.generateKsonnet(options)
-		if ksonnetErr != nil {
-			return fmt.Errorf("could not generate kssonnet under %v Error: %v", kstypes.KsName, ksonnetErr)
-		}
-	case kftypes.ALL:
-		gcpConfigFilesErr := gcp.generateDMConfigs(options)
-		if gcpConfigFilesErr != nil {
-			return fmt.Errorf("could not generate deployment manager configs under %v Error: %v", GCP_CONFIG, gcpConfigFilesErr)
-		}
-		generateK8sSpecsErr := gcp.downloadK8sManifests()
-		if generateK8sSpecsErr != nil {
-			return fmt.Errorf("could not generate files under %v Error: %v", K8S_SPECS, generateK8sSpecsErr)
-		}
-		ksonnetErr := gcp.generateKsonnet(options)
-		if ksonnetErr != nil {
-			return fmt.Errorf("could not generate kssonnet under %v Error: %v", kstypes.KsName, ksonnetErr)
-		}
-	case kftypes.PLATFORM:
-		gcpConfigFilesErr := gcp.generateDMConfigs(options)
-		if gcpConfigFilesErr != nil {
-			return fmt.Errorf("could not generate deployment manager configs under %v Error: %v", GCP_CONFIG, gcpConfigFilesErr)
-		}
-		ksonnetErr := gcp.generateKsonnet(options)
-		if ksonnetErr != nil {
-			return fmt.Errorf("could not generate kssonnet under %v Error: %v", kstypes.KsName, ksonnetErr)
-		}
-	}
-	createConfigErr := gcp.writeConfigFile()
-	if createConfigErr != nil {
-		return fmt.Errorf("cannot create config file app.yaml in %v", gcp.GcpApp.Spec.AppDir)
-	}
-	return nil
-}
-
 func (gcp *Gcp) getServiceClient(ctx context.Context) (*http.Client, error) {
 
 	// See https://cloud.google.com/docs/authentication/.
@@ -894,65 +429,4 @@ func (gcp *Gcp) getServiceClient(ctx context.Context) (*http.Client, error) {
 		return nil, err
 	}
 	return client, nil
-}
-
-func (gcp *Gcp) gcpInitProject() error {
-	ctx := context.Background()
-	client, clientErr := gcp.getServiceClient(ctx)
-	if clientErr != nil {
-		return fmt.Errorf("could not create client %v", clientErr)
-	}
-	serviceusageService, serviceusageServiceErr := serviceusage.New(client)
-	if serviceusageServiceErr != nil {
-		return fmt.Errorf("could not create service usage service %v", serviceusageServiceErr)
-	}
-
-	enabledApis := []string{
-		"deploymentmanager.googleapis.com",
-		"servicemanagement.googleapis.com",
-		"container.googleapis.com",
-		"cloudresourcemanager.googleapis.com",
-		"endpoints.googleapis.com",
-		"file.googleapis.com",
-		"ml.googleapis.com",
-		"iam.googleapis.com",
-		"sqladmin.googleapis.com",
-	}
-	for _, api := range enabledApis {
-		service := fmt.Sprintf("projects/%v/services/%v", gcp.GcpApp.Spec.Project, api)
-		_, opErr := serviceusageService.Services.Enable(service, &serviceusage.EnableServiceRequest{}).Context(ctx).Do()
-		if opErr != nil {
-			return fmt.Errorf("could not enable API service %v: %v", api, opErr)
-		}
-	}
-	return nil
-}
-
-func (gcp *Gcp) Init(options map[string]interface{}) error {
-	ks := gcp.Children[kftypes.KSONNET]
-	if ks != nil {
-		ksInitErr := ks.Init(options)
-		if ksInitErr != nil {
-			return fmt.Errorf("gcp init failed for %v: %v", string(kftypes.KSONNET), ksInitErr)
-		}
-	} else {
-		return fmt.Errorf("%v not in Children", string(kftypes.KSONNET))
-	}
-	cacheDir := path.Join(gcp.GcpApp.Spec.AppDir, kftypes.DefaultCacheDir)
-	newPath := filepath.Join(cacheDir, gcp.GcpApp.Spec.Version)
-	gcp.GcpApp.Spec.Repo = path.Join(newPath, "kubeflow")
-	createConfigErr := gcp.writeConfigFile()
-	if createConfigErr != nil {
-		return fmt.Errorf("cannot create config file app.yaml in %v", gcp.GcpApp.Spec.AppDir)
-	}
-
-	if !gcp.GcpApp.Spec.SkipInitProject {
-		log.Infof("Not skipping GCP project init, running gcpInitProject.")
-		initProjectErr := gcp.gcpInitProject()
-		if initProjectErr != nil {
-			return fmt.Errorf("cannot init gcp project %v", initProjectErr)
-		}
-	}
-
-	return nil
 }
